requirement.txt
asgiref==3.3.4
bleach==3.3.0
Django==3.2.4
django-crispy-forms==1.12.0
django-markdownify==0.9.0
django-qr-code==2.2.0
djangorestframework==3.12.4
install==1.3.4
Markdown==3.3.4
packaging==21.0
Pillow==8.2.0
Pygments==2.9.0
pyparsing==2.4.7
pytz==2021.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
webencodings==0.5.1
